refactor: restructure source code for better organization
