+++ ui/app/strategy/[id]/components/strategy-card.tsx
