+++ ui/app/strategy/[id]/components/strategy-flow.tsx
