+++ ui/app/strategy/[id]/components/strategy-header.tsx
