+++ ui/app/strategy/[id]/components/strategy-overview.tsx
