+++ ui/app/strategy/[id]/components/strategy-tabs.tsx
