+++ ui/components/shared/wallet-connect-modal.tsx
